# src/main/java/ai/toby/reminder/controller/request/UpdateReminderListRequest.java
package ai.toby.reminder.controller.request;

import jakarta.validation.constraints.NotBlank;

public record UpdateReminderListRequest(
        @NotBlank String name,
        String color,
        String icon
) {}